requirements.txt
anyio==3.5.0
appdirs==1.4.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asks==2.3.6
asttokens==2.0.5
async-exit-stack==1.0.1
async-generator==1.10
atomicwrites==1.4.0
attrs==19.3.0
autopep8==1.6.0
Babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
blinker==1.5
blis==0.7.8
boto3==1.24.77
botocore==1.27.77
bs4==0.0.1
bump2version==0.5.11
cachetools==5.0.0
catalogue==2.0.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
Click==7.0
colorama==0.4.4
common==0.1.2
commonmark==0.9.1
crispy-bootstrap5==0.6
cryptography==37.0.4
cycler==0.11.0
cymem==2.0.6
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.4
Django==4.1.1
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-environ==0.9.0
django-filter==22.1
django-rest-knox==4.2.0
django-rest-passwordreset==1.2.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
docker==6.0.0
docutils==0.14
dynamorm==0.11.0
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
filelock==3.6.0
flashtext==2.7
Flask==2.1.2
fonttools==4.33.3
for==0.0.0.0.0.0a0.dev20210120
future==0.18.2
greenlet==1.1.2
h11==0.13.0
helpers==0.2.0
huggingface-hub==0.7.0
idna==3.3
importlib-metadata==0.23
iniconfig==1.1.1
ipykernel==6.13.0
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
itsdangerous==2.1.2
jedi==0.18.1
jellyfish==0.9.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.1.0
json5==0.9.8
jsonschema==4.5.1
jupyter==1.0.0
jupyter-client==7.3.1
jupyter-console==6.4.3
jupyter-core==4.10.0
jupyter-server==1.17.1
jupyterlab==3.4.3
jupyterlab-pygments==0.2.2
jupyterlab-server==2.14.0
jupyterlab-widgets==1.1.0
kiwisolver==1.4.2
langcodes==3.3.0
lief==0.12.1
MarkupSafe==2.1.1
marshmallow==3.2.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
migrations==0.1.1
mistune==0.8.4
more-itertools==8.12.0
MouseInfo==0.1.3
murmurhash==1.0.7
mypy==0.942
mypy-extensions==0.4.3
mysqlclient==2.1.1
nbclassic==0.3.7
nbclient==0.6.3
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.8.2
nltk==3.7
notebook==6.4.11
notebook-shim==0.1.0
numpy==1.23.2
outcome==1.1.0
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.5.1
pathy==0.6.2
pefile==2021.9.3
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
pluggy==0.13.1
preshed==3.0.6
prometheus-client==0.14.1
prompt-toolkit==3.0.29
psutil==5.9.1
psycopg2==2.9.3
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.2
pygame==2.1.2
PyGetWindow==0.0.9
Pygments==2.12.0
PyJWT==2.4.0
PyMsgBox==1.0.9
pyparsing==3.0.7
pyperclip==1.8.2
PyRect==0.2.0
pyrsistent==0.14.11
PyScreeze==0.1.28
PySocks==1.7.1
pytest==5.4.3
python-dateutil==2.8.2
python-decouple==3.6
pytube==12.0.0
pytweening==1.0.4
pytz==2022.2.1
pywin32==304
pywinpty==2.0.5
PyYAML==5.1.1
pyzmq==23.0.0
qtconsole==5.3.0
QtPy==2.1.0
rake-nltk==1.0.6
regex==2022.4.24
requests==2.27.1
rich==12.4.4
rsa==4.8
s3transfer==0.6.0
scikit-learn==1.1.1
scipy==1.8.1
segtok==1.5.11
selenium==4.4.3
Send2Trash==1.8.0
sentencepiece==0.1.96
six==1.16.0
sklearn==0.0
smart-open==5.2.1
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spacy==3.3.1
spacy-legacy==3.0.9
spacy-loggers==1.0.2
SQLAlchemy==1.4.39
sqlparse==0.4.2
srsly==2.4.3
stack-data==0.2.0
tabulate==0.8.9
terminado==0.15.0
thinc==8.0.17
threadpoolctl==3.1.0
tinycss2==1.1.1
tk==0.1.0
tokenizers==0.12.1
toml==0.10.0
tomli==2.0.1
toolz==0.10.0
torch==1.11.0
torchvision==0.12.0
tornado==6.1
tqdm==4.64.0
traitlets==5.2.1.post0
transformers==4.19.2
trio==0.21.0
trio-typing==0.2.0
trio-websocket==0.9.2
twilio==7.12.1
typer==0.4.2
typing_extensions==4.1.1
tzdata==2022.2
urllib3==1.26.9
vector==0.8.5
virtualenv==20.14.1
wasabi==0.9.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.3
Werkzeug==2.1.2
widgetsnbextension==3.6.0
wsproto==1.2.0
xmltodict==0.12.0
yake==0.4.8
zipp==3.7.0
